requirements.txt
# AI/ML Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.20.0
safetensors>=0.3.0

# Stable Diffusion and ControlNet
controlnet-aux>=0.0.6
xformers>=0.0.20

# Image Processing
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0
scipy>=1.10.0

# 3D and CAD Processing
trimesh>=3.20.0
open3d>=0.17.0
pyvista>=0.40.0

# Web Interface
streamlit>=1.25.0
gradio>=3.40.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Optional: GPU acceleration
# nvidia-ml-py3>=7.352.0
# cupy-cuda11x>=12.0.0  # For CUDA 11.x
# cupy-cuda12x>=12.0.0  # For CUDA 12.x
